\name{AtomicWeight}
\docType{data}
\alias{AtomicWeight}
\title{The atomic weights of chemical elements}
\description{Atomic weights of chemical elements useful for marine and lacustrine sciences}
\usage{AtomicWeight}
\format{
  a dataframe
}
\author{Karline Soetaert <k.soetaert@nioo.knaw.nl>}
\examples{AtomicWeight
AtomicWeight$H
(W_H2O<- with (AtomicWeight, 2*H + O))
}
\references{to be filled...}
\keyword{datasets}


