\name{sal2cl}
\alias{sal2cl}
\title{Salinity-chlorinity conversion}
\description{Estimates the chlorinity concentration based on salinity}
\usage{sal2cl(S=35)}
\arguments{
  \item{S }{salinity}
}
\value{
  The chlorinity concentration, g/kg
}
\author{Karline Soetaert <k.soetaert@nioo.knaw.nl>}
\examples{sal2cl(20)}
\references{Wooster et al.,1969}
\keyword{utilities}

